refactor: MAT-269 교체선수 리스트 컴포넌트 이름을 SimplePlayerList로 변경
