+++ src/components/SimplePlayerList/SimplePlayerListContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './SimplePlayerListContainer';
