+++ src/components/SimplePlayerList/SimplePlayerListItem/SimplePlayerListItem.css.ts
