+++ src/components/SimplePlayerList/SimplePlayerListItem/index.ts
@@ -1,0 +1,1 @@
+export * from './SimplePlayerListItem';
