--- src/components/SubstitutionPlayerList/SubstitutionPlayerListContainer/index.ts
@@ -1,1 +1,0 @@
-export * from './SubstitutionPlayerListContainer';
