--- src/components/SubstitutionPlayerList/SubstitutionPlayerListItem/index.ts
@@ -1,1 +1,0 @@
-export * from './SubstitutionPlayerListItem';
