--- src/components/SubstitutionPlayerList/index.ts
@@ -1,1 +1,0 @@
-export * from './SubstitutionPlayerList';
